requirements.txt
# Core async + HTTP
aiohttp>=3.9.0
httpx>=0.27.0
requests>=2.32.0

# LLM providers
openai>=1.30.0
anthropic>=0.28.0
google-genai>=1.0.0
google-generativeai>=0.5.0
ollama>=0.2.0

# CLI & display
rich>=13.7.0
typer>=0.12.0
colorama>=0.4.6

# Data & config
pydantic>=2.7.0
python-dotenv>=1.0.0
tabulate>=0.9.0

# DNS
dnspython>=2.6.0

# Graph generation
networkx>=3.3
pyvis>=0.3.2

# Web parsing
beautifulsoup4>=4.12.0
lxml>=5.2.0

# JS analysis
jsbeautifier>=1.15.0

# Cloud
boto3>=1.34.0

# Secret scanning (Python wrappers / tools)
# Note: trufflehog and gitleaks are Go binaries — install separately
# pip install cloud-enum           # CloudEnum
# pip install dnsgen               # dnsgen  
# pip install altdns               # altdns
# pip install xnLinkFinder         # xnLinkFinder
# pip install linkfinder           # LinkFinder

# Template / report
jinja2>=3.1.4
tqdm>=4.66.0
